+++ src/extension.ts
@@ -33,7 +33,11 @@ export function activate(context: vscode.ExtensionContext): void {
   const normalizedViewProvider = new NormalizedViewContentProvider();
   const compareViewProvider = new CompareViewContentProvider();
   const mergedViewProvider = new MergedViewContentProvider(context.globalStorageUri);
-  const interactiveViewController = new InteractiveViewPanelController(context.extensionUri);
+  const interactiveViewController = new InteractiveViewPanelController(
+    context.extensionUri,
+    normalizedViewProvider,
+    mergedViewProvider
+  );
 
   context.subscriptions.push(
     vscode.workspace.registerTextDocumentContentProvider(